# [Java] Alignment checks and optimizations.