--- a/agrona-concurrency-tests/src/main/java/org/agrona/concurrent/ringbuffer/ManyToOneRingBufferTests.java
+++ b/agrona-concurrency-tests/src/main/java/org/agrona/concurrent/ringbuffer/ManyToOneRingBufferTests.java
@@ -18,13 +18,19 @@ package org.agrona.concurrent.ringbuffer;
 import org.agrona.DirectBuffer;
 import org.agrona.ExpandableArrayBuffer;
 import org.agrona.concurrent.UnsafeBuffer;
-import org.openjdk.jcstress.annotations.*;
+import org.openjdk.jcstress.annotations.Actor;
+import org.openjdk.jcstress.annotations.Arbiter;
+import org.openjdk.jcstress.annotations.Expect;
+import org.openjdk.jcstress.annotations.JCStressMeta;
+import org.openjdk.jcstress.annotations.JCStressTest;
+import org.openjdk.jcstress.annotations.Outcome;
+import org.openjdk.jcstress.annotations.State;
 import org.openjdk.jcstress.infra.results.II_Result;
 import org.openjdk.jcstress.infra.results.JJJ_Result;
 
+import static java.nio.ByteBuffer.allocateDirect;
 import static org.agrona.BitUtil.SIZE_OF_INT;
 import static org.agrona.BitUtil.SIZE_OF_LONG;
-import static org.agrona.BufferUtil.allocateDirectAligned;
 import static org.agrona.concurrent.ringbuffer.ManyToOneRingBuffer.MIN_CAPACITY;
 import static org.agrona.concurrent.ringbuffer.RingBufferDescriptor.TRAILER_LENGTH;
 
@@ -55,7 +61,7 @@ public class ManyToOneRingBufferTests
     public static class Write
     {
         private static final int MSG_TYPE_ID = 7;
-        private final ManyToOneRingBuffer ringBuffer = new ManyToOneRingBuffer(new UnsafeBuffer(new long[128]));
+        private final ManyToOneRingBuffer ringBuffer = new ManyToOneRingBuffer(new UnsafeBuffer(allocateDirect(1024)));
         private final ExpandableArrayBuffer srcBuffer = new ExpandableArrayBuffer();
 
         /**
@@ -118,7 +124,7 @@ public class ManyToOneRingBufferTests
     {
         private static final int MSG_TYPE_ID = 11;
         private final ManyToOneRingBuffer ringBuffer =
-            new ManyToOneRingBuffer(new UnsafeBuffer(allocateDirectAligned(1024, SIZE_OF_LONG)));
+            new ManyToOneRingBuffer(new UnsafeBuffer(new UnsafeBuffer(allocateDirect(1024))));
 
         /**
          * First writer thread.
@@ -174,7 +180,7 @@ public class ManyToOneRingBufferTests
     public static class TryClaimAbort
     {
         private static final int MSG_TYPE_ID = 19;
-        private final ManyToOneRingBuffer ringBuffer = new ManyToOneRingBuffer(new UnsafeBuffer(new long[128]));
+        private final ManyToOneRingBuffer ringBuffer = new ManyToOneRingBuffer(new UnsafeBuffer(allocateDirect(1024)));
 
         /**
          * First writer thread.
@@ -239,7 +245,7 @@ public class ManyToOneRingBufferTests
     public static class CorrelationId
     {
         private final ManyToOneRingBuffer ringBuffer =
-            new ManyToOneRingBuffer(new UnsafeBuffer(new long[(MIN_CAPACITY + TRAILER_LENGTH) / SIZE_OF_LONG]));
+            new ManyToOneRingBuffer(new UnsafeBuffer(allocateDirect(MIN_CAPACITY + TRAILER_LENGTH)));
 
         /**
          * First thread.